Add a g:ruby_no_hanging_indent option
Setting this to 1 will change the indentation behavior so that
pressing Return and adding a ) after the following code

    invoice = FactoryGirl.create(:invoice,

no longer aligns the ) with the (, but places it at the same level as the above
line. A similar thing will happen for

    some_long_variable = case

instead of indenting two spaces after the beginning of the `case`.